finish add express validator

+++ controllers/user/delete.js
@@ -1,16 +1,27 @@
+const { validationResult } = require("express-validator");
 const saveRedis = require("../../helper/saveRedis");
 const UserData = require("../../models")
 const { ObjectId } = require('bson');
+const { generateValidatorError } = require("../../helper/generateErrors");
 
 async function deleteUser
     (req, res) {
     try {
 
+        const errors = validationResult(req);
+        if (!errors.isEmpty()) throw (generateValidatorError(errors.array()));
+
         const { user_id } = req.body
 
-        await UserData.deleteOne(
+        const resultDelete = await UserData.deleteOne(
             { _id: new ObjectId(user_id) },
         );
+
+        if (resultDelete.deletedCount !== 1) {
+            throw {
+                message: "user is failed to delete"
+            }
+        }
 
         await saveRedis()
